Refactor main logic into modular package
