+++ borderframe/__init__.py
@@ -1,0 +1,7 @@
+"""BorderFrame application package."""
+
+from .image_processor import ImageProcessor
+from .thumbnail_dialog import ThumbnailDialog
+from .process_worker import ProcessWorker
+
+__all__ = ["ImageProcessor", "ThumbnailDialog", "ProcessWorker"]
